feat: add support for slovenian language
Refs: SHELL-294
