+++ src/settings/general-settings.test.tsx
@@ -9,9 +9,9 @@ import 'jest-styled-components';
 import { screen, waitFor, within } from '@testing-library/react';
 import { find } from 'lodash';
 
-import type { LocaleDescriptorWithLabels } from './components/utils';
-import { localeList } from './components/utils';
 import GeneralSettings from './general-settings';
+import type { LocaleDescriptorWithLabels } from '../constants/locales';
+import { localeList } from '../constants/locales';
 import { useAccountStore } from '../store/account';
 import { useI18nStore } from '../store/i18n/store';
 import { useLoginConfigStore } from '../store/login/store';
